backend/requirements.txt
fastapi
uvicorn
httpx
pytest
neontology
celery[redis]
pyjwt
passlib
bcrypt==4.0.1
pydantic-settings
watchfiles
watchdog
pycryptodome
python-jose[cryptography]
redis
cryptography
requests
impacket
reportlab